fixed push constant error for copy kernel

+++ src/core/slice.cpp
@@ -43,13 +43,13 @@ namespace core {
         DLPRIM_CHECK(s[2] == t[2]);
         int p = 0;
 #if VULKAN_API
-		kernel_->setArg(p++,uint64_t(slice));
-        kernel_->setArg(p++,uint64_t(s[0]));
-        kernel_->setArg(p++,uint64_t(t[1]));
-        kernel_->setArg(p++,uint64_t(target_offset));
-        kernel_->setArg(p++,uint64_t(s[1]));
-        kernel_->setArg(p++,uint64_t(source_offset));
-        kernel_->setArg(p++,uint64_t(s[2]));
+		kernel_->setArg(p++,uint32_t(slice));
+        kernel_->setArg(p++,uint32_t(s[0]));
+        kernel_->setArg(p++,uint32_t(t[1]));
+        kernel_->setArg(p++,uint32_t(target_offset));
+        kernel_->setArg(p++,uint32_t(s[1]));
+        kernel_->setArg(p++,uint32_t(source_offset));
+        kernel_->setArg(p++,uint32_t(s[2]));
         target.set_arg(kernel_,p);
         source.set_arg(kernel_,p);
         bind_as_dtype(kernel_,p,scale,dtype_);
